requirements.txt
openai>=1.50.0
httpx==0.27.0
requests==2.32.3
python-dotenv==1.0.1
sounddevice==0.4.7
scipy==1.13.1
numpy==1.26.4
soundfile